@@ -13,13 +13,15 @@
     "autoprefixer": "^10.4.7",
     "clsx": "^1.1.1",
     "dayjs": "^1.11.3",
+    "eslint-config-next": "^13.3.4",
     "framer-motion": "^6.5.1",
-    "next": "^12.1.6",
+    "next": "^13.3.4",
     "postcss": "^8.4.14",
     "prop-types": "^15.8.1",
-    "react": "^18.1.0",
-    "react-dom": "^18.1.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-icons": "^4.8.0",
+    "server-only": "^0.0.1",
     "tailwindcss": "^3.1.0"
   },
   "devDependencies": {
